feat: Add utility functions for message rendering and emotion dimension snapshot
