+++ ims_bot_adapter/src/lib.rs
@@ -21,6 +21,7 @@ pub mod send_message;
 pub mod send_qq_message_batches;
 pub mod system_config;
 pub mod tools;
+pub mod utils;
 pub mod ws_action;
 
 use zihuan_core::error::Result;
