+++ zihuan_agent/src/emotion/utils.rs
@@ -1,0 +1,22 @@
+use crate::session_state::QqChatAgentSessionState;
+use zihuan_core::agent_config::QqChatEmotionDimensionConfig;
+
+/// Formats the current emotion dimension state as a multi-line display string.
+///
+/// Each line follows the pattern: `{name}: {value}`.
+/// Returns `[No emotion dimensions]` when no dimensions are configured.
+pub fn emotion_dimensions_snapshot_text(
+    session_state: &QqChatAgentSessionState,
+    emotion_dimensions: &[QqChatEmotionDimensionConfig],
+) -> String {
+    let lines: Vec<String> = session_state
+        .ordered_emotion_dimensions(emotion_dimensions)
+        .into_iter()
+        .map(|(name, value)| format!("{name}: {value}"))
+        .collect();
+    if lines.is_empty() {
+        "[No emotion dimensions]".to_string()
+    } else {
+        lines.join("\n")
+    }
+}
