+++ zihuan_agent/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod brain;
+pub mod emotion;
 pub mod session_state;
 pub mod utils;
